+++ code/src/cljs/sixsq/nuvla/ui/dashboard/spec.cljs
@@ -1,4 +1,4 @@
-(ns sixsq.nuvla.ui.deployment.spec
+(ns sixsq.nuvla.ui.dashboard.spec
   (:require-macros [sixsq.nuvla.ui.utils.spec :refer [only-keys]])
   (:require
     [clojure.spec.alpha :as s]))